requirements.txt
aiofiles==25.1.0
amqp==5.3.1
anyio==4.12.0
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.12.2
billiard==4.2.4
blinker==1.9.0
celery==5.6.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.68.0
Flask==2.3.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
kombu==5.6.1
MarkupSafe==3.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.7
numpy==2.3.5
openai==2.9.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
protobuf==4.21.12
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==1.10.24
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
reportlab==4.4.5
requests==2.31.0
selenium==4.15.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.4
starlette==0.14.2
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.15.0
vine==5.1.0
wcwidth==0.2.14
webdriver-manager==4.0.2
Werkzeug==2.3.7
whitenoise==6.11.0
wsproto==1.2.0
